POC, move the code of e2e tests into a built file.
